Move api.js into lib/
Originally the goal may have been for this to be accessible
programmatically, but it's unclear whether we'd ever end up supporting
that. For now moving it into the lib/ directory better communicates it's
internal.
