+++ lib/api.js
@@ -14,12 +14,12 @@ const arrify = require('arrify');
 const makeDir = require('make-dir');
 const ms = require('ms');
 const chunkd = require('chunkd');
-const babelPipeline = require('./lib/babel-pipeline');
-const Emittery = require('./lib/emittery');
-const RunStatus = require('./lib/run-status');
-const AvaFiles = require('./lib/ava-files');
-const fork = require('./lib/fork');
-const serializeError = require('./lib/serialize-error');
+const babelPipeline = require('./babel-pipeline');
+const Emittery = require('./emittery');
+const RunStatus = require('./run-status');
+const AvaFiles = require('./ava-files');
+const fork = require('./fork');
+const serializeError = require('./serialize-error');
 
 function resolveModules(modules) {
 	return arrify(modules).map(name => {
