+++ lib/cli.js
@@ -156,7 +156,7 @@ exports.run = () => { // eslint-disable-line complexity
 	}
 
 	const ciParallelVars = require('ci-parallel-vars');
-	const Api = require('../api');
+	const Api = require('./api');
 	const VerboseReporter = require('./reporters/verbose');
 	const MiniReporter = require('./reporters/mini');
 	const TapReporter = require('./reporters/tap');
